tipc: Incorporate scope information into name publication distribution
Enhances TIPC's name distribution messages to include information about
the publication scope of the associated name. This allows the nodes
receiving the publication message to properly display scope information
about non-local name table entries.

Scope information is now encoded in the low 2 bits of the publication key,
with the remaining bits representing a count of the number of publications
the port has issued. The low order bit of a new "flags" field in the header
of name distribution messages now indicates if this encoding is being used,
allowing TIPC to assign a scope of 0 (meaning "unknown") to publications
issued by nodes that pre-date this capability; such name table entries are
displayed using a blank scope field.

The revised encoding of the publication key has no impact on nodes running
earlier versions of TIPC, as they do not attempt to attribute meaning to
the individual bits of the key.

Signed-off-by: Allan Stephens <allan.stephens@windriver.com>
Signed-off-by: Erik Hugne <erik.hugne@ericsson.com>