tipc: Don't track sent message count for connectionless sockets
Eliminates code that tracks the number of sent messages for connectionless
sockets, since this value is only needed by the flow control algorithm for
connection-based sockets. The one instance in which a connection-based
socket utilizes the revised port-layer send routines now increments the
counter in a socket-layer send routine to ensure that connection-based
message tracking is not affected.

Signed-off-by: Ying Xue <ying.xue@windriver.com>
Signed-off-by: Allan Stephens <allan.stephens@windriver.com>
Signed-off-by: Erik Hugne <erik.hugne@ericsson.com>