tipc: Clarify message-based flow control for connected sockets
Replaces the existing symbol used by TIPC's connection-based flow
control algorithm with a pair of symbols: one which denotes the maximum
number of unacknowledged messages a sending socket may emit before
prohibiting further sends and one which denotes the maximum number of
unacknowledged messages a receiving socket may process before sending
acknowledgement information to the sending socket. These new symbols
are designed to make it clear that the latter value is 50% of the former,
which ensures that the sender receives acknowledgements often enough
to prevent it from blocking unnecessarily.

Note that this change is essentially cosmetic in nature, and does not
alter the operation of TIPC.

Signed-off-by: Ying Xue <ying.xue@windriver.com>
Signed-off-by: Allan Stephens <allan.stephens@windriver.com>
Signed-off-by: Erik Hugne <erik.hugne@ericsson.com>